# src/pkg/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "The IPEHR Gateway is an openEHR compliant EHR server implementation that stores encrypted medical data in a Filecoin distributed file storage.",
        "title": "IPEHR Gateway API",
        "contact": {
            "name": "API Support",
            "url": "https://bsn.si/blockchain",
            "email": "support@bsn.si"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.2"
    },
    "host": "gateway.ipehr.org",
    "basePath": "/v1",
    "paths": {
...
        "/requests/": {
            "get": {
                "description": "It is returning only transactions which in progress\n",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "REQUEST"
                ],
                "summary": "Get list of transactions requests by authorized user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer AccessToken",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "UserId",
                        "name": "AuthUserId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "default: 10",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "id namespace. Example: examples",
                        "name": "offset",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/processing.RequestsResult"
                        }
                    },
                    "400": {
                        "description": "Is returned when userID is empty"
                    },
                    "404": {
                        "description": "Is returned when requests not exist"
                    },
                    "500": {
                        "description": "Is returned when an unexpected error occurs while processing a request"
                    }
                }
            }
        },
        "/requests/{request_id}": {
            "get": {
                "description": "It's returning only transactions which in progress\n",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "REQUEST"
                ],
                "summary": "Get list of transactions by certain request id for authorized user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer AccessToken",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "UserId",
                        "name": "AuthUserId",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Unique id of request",
                        "name": "request_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/processing.RequestResult"
                        }
                    },
                    "400": {
                        "description": "Is returned when userID or request_id is empty"
                    },
                    "404": {
                        "description": "Is returned when requests not exist"
                    },
                    "500": {
                        "description": "Is returned when an unexpected error occurs while processing a request"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "processing.EthereumTx": {
            "type": "object",
            "properties": {
                "Kind": {
                    "type": "string"
                },
                "Status": {
                    "type": "string"
                },
                "comment": {
                    "type": "string"
                },
                "hash": {
                    "type": "string"
                }
            }
        },
        "processing.FileCoinTx": {
            "type": "object",
            "properties": {
                "Kind": {
                    "type": "string"
                },
                "Status": {
                    "type": "string"
                },
                "cid": {
                    "type": "string"
                },
                "comment": {
                    "type": "string"
                },
                "dealCID": {
                    "type": "string"
                },
                "dealID": {
                    "type": "integer"
                },
                "minerAddress": {
                    "type": "string"
                }
            }
        },
        "processing.RequestResult": {
            "type": "object",
            "properties": {
                "ethereum": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/processing.EthereumTx"
                    }
                },
                "filecoin": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/processing.FileCoinTx"
                    }
                },
                "kind": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "processing.RequestsResult": {
            "type": "object",
            "additionalProperties": {
                "$ref": "#/definitions/processing.RequestResult"
            }
        }
    }
}